+++ test/application_test.exs
@@ -5,4 +5,8 @@ defmodule Money.Application.Test do
     {_app, options} = Application.spec(:ex_money) |> Keyword.get(:mod)
     assert options == [strategy: :one_for_one, name: Money.Supervisor]
   end
+
+  test "default supervisor name" do
+    assert Money.ExchangeRates.Supervisor.default_supervisor == Money.Supervisor
+  end
 end